src/app/utils/s3.py
import boto3
import os


def get_s3_client():
    return boto3.client(
        "s3",
        endpoint_url=os.getenv("MINIO_ENDPOINT"),
        aws_access_key_id=os.getenv("MINIO_USER"),
        aws_secret_access_key=os.getenv("MINIO_PASSWORD"),
        config=boto3.session.Config(signature_version="s3v4"),
        region_name="us-east-1"
    )