pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "analytics"
version = "0.1.0"
dependencies = [
    "mysql-connector-python",
    "python-dotenv",
    "pandas",
    "sqlalchemy",
    "pyspark==3.5.1"

]

[tool.setuptools.packages.find]
where = ["src"]
include = ["app*"]
